+++ Examples/2.x/Example/Modules/Images/Sources/Images/ContentView.swift
@@ -29,7 +29,8 @@ public struct ContentView: View {
             // SwiftGen ussing:
             // Asset.Folder.duplicatedImage3.image
             // Asset.duplicatedImage1.image
-            // Asset.snakeCaseImage.image
+            // Asset.snakeCaseImage.name
+            
             Image("snake-case_image")
         }
         .padding()